Update the structure based on the events refactor
For more context, see
https://github.com/district09/robo-digipolis-helpers/commit/6b756a6e4f120b7fcb149f389d2d61334a4fee51

This update is done to be compatible with the upcoming version of the
helpers base package. Why it is needed is explained in the commit linked
above.
